requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
SpeechRecognition==3.8.1
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.11.0
gTTS==2.4.0
requests>=2.31.0
python-dateutil>=2.8.2
PyAudio==0.2.13
pygame>=2.5.2 